Change foreign datatype

+++ database/migrations/2019_05_30_200929_create_multi_images_table.php
@@ -15,7 +15,7 @@ class CreateMultiImagesTable extends Migration
     {
         Schema::create('multi_images', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('product_id');
+            $table->unsignedTinyInteger('product_id');
             $table->string('product_image');
             $table->timestamps();
             //$table->foreign('product_id')->references('id')->on('products')->onUpdate('cascade')->onDelete('cascade');
